separo los metodos en distintas clases. corriijo mayusculas. corrijo algunas entidades. creo el menu principal pero me da error.
