+++ src/main/java/um/edu/uy/Exceptions/ElementAlreadyExists.java
@@ -1,0 +1,7 @@
+package um.edu.uy.Exceptions;
+
+public class ElementAlreadyExists extends Exception {
+    public ElementAlreadyExists(String message) {
+        super(message);
+    }
+}
