+++ src/main/java/um/edu/uy/Exceptions/ElementNotFound.java
@@ -1,0 +1,7 @@
+package um.edu.uy.Exceptions;
+
+public class ElementNotFound extends Exception {
+    public ElementNotFound(String message) {
+        super(message);
+    }
+}
