+++ src/main/java/um/edu/uy/Exceptions/EmptyHeapException.java
@@ -1,0 +1,7 @@
+package um.edu.uy.Exceptions;
+
+public class EmptyHeapException extends Exception{
+    public EmptyHeapException(String mensaje ){
+        System.out.println(mensaje);
+    }
+}
