+++ src/main/java/um/edu/uy/Exceptions/InvalidIndex.java
@@ -1,0 +1,7 @@
+package um.edu.uy.Exceptions;
+
+public class InvalidIndex extends Exception {
+    public InvalidIndex(String message) {
+        super(message);
+    }
+}
