+++ src/main/java/um/edu/uy/Exceptions/ValueNoExists.java
@@ -1,0 +1,7 @@
+package um.edu.uy.Exceptions;
+
+public class ValueNoExists extends Exception {
+    public ValueNoExists(String message) {
+        super(message);
+    }
+}
