--- src/main/java/um/edu/uy/General/CargarActoresYDirectores.java
@@ -1,4 +1,0 @@
-package um.edu.uy.General;
-
-public class CargarActoresYDirectores {
-}
