+++ src/main/java/um/edu/uy/Metodos/UsuarioConMasCalificacionesPorGenero.java
@@ -1,0 +1,4 @@
+package um.edu.uy.Metodos;
+
+public class UsuarioConMasCalificacionesPorGenero {
+}
