--- src/main/java/um/edu/uy/entities/Actor.java
@@ -1,19 +1,0 @@
-package um.edu.uy.entities;
-
-public class Actor {
-    private String nombre;
-
-    //constructor
-    public Actor(String nombre) {
-        this.nombre = nombre;
-    }
-
-    //getter y setter
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-}
