--- src/main/java/um/edu/uy/exceptions/ElementAlreadyExists.java
@@ -1,7 +1,0 @@
-package um.edu.uy.exceptions;
-
-public class ElementAlreadyExists extends Exception {
-    public ElementAlreadyExists(String message) {
-        super(message);
-    }
-}
