--- src/main/java/um/edu/uy/exceptions/ElementNotFound.java
@@ -1,7 +1,0 @@
-package um.edu.uy.exceptions;
-
-public class ElementNotFound extends Exception {
-    public ElementNotFound(String message) {
-        super(message);
-    }
-}
