--- src/main/java/um/edu/uy/exceptions/EmptyHeapException.java
@@ -1,7 +1,0 @@
-package um.edu.uy.exceptions;
-
-public class EmptyHeapException extends Exception{
-    public EmptyHeapException(String mensaje ){
-        System.out.println(mensaje);
-    }
-}
