--- src/main/java/um/edu/uy/exceptions/InvalidIndex.java
@@ -1,7 +1,0 @@
-package um.edu.uy.exceptions;
-
-public class InvalidIndex extends Exception {
-    public InvalidIndex(String message) {
-        super(message);
-    }
-}
