--- src/main/java/um/edu/uy/exceptions/ValueNoExists.java
@@ -1,7 +1,0 @@
-package um.edu.uy.exceptions;
-
-public class ValueNoExists extends Exception {
-    public ValueNoExists(String message) {
-        super(message);
-    }
-}
